nav2_rviz_plugins: Remove slots without implementation (#4974)
Without this change, when compiled for Jazzy with Nix using gcc14 and
qt-5.15, loading the plugins to rviz fails with the following error:

    [rviz2]: PluginlibFactory: The plugin for class
    'nav2_rviz_plugins/Navigation 2' failed to load. Error: Failed to
    load library /nix/store/...-ros-env/lib/libnav2_rviz_plugins.so.
    Make sure that you are calling the PLUGINLIB_EXPORT_CLASS macro in
    the library code, and that names are consistent between this macro
    and your XML. Error string: Could not load library dlopen error:
    /nix/store/...-ros-env/lib/libnav2_rviz_plugins.so: undefined
    symbol: _ZN17nav2_rviz_plugins12DockingPanel9onStartupEv, at
    /build/rcutils-release-release-jazzy-rcutils-6.7.2-1/src/shared_library.c:99

The dynamic linker complains about undefined symbol, which corresponds
to a Qt slot, which is declared, but has no implementation. I'm not
sure what is the root cause of this behavior, but removing the slot
definitions fixes the problem and the plugins can be loaded.

Signed-off-by: Michal Sojka <michal.sojka@cvut.cz>
Signed-off-by: Sakshay Mahna <sakshum19@gmail.com>